Minor layout edits
While design considerations are ongoing, this commit updates the
directory structure of 'optimacore' to adhere to standards.
The README.md is empty but available for documentation.
The CHANGELOG.md file is an attempt at version documentation.
The setup.py, __init__.py and _version.py trio of setup files have also
been introduced, with the former two heavily commented out for the time
being.