Logging information improved
Portfolio and Project now have a 'name' attribute, intended to be their
human readable IDs used for reference.
These are accompanied by set and get methods for good practice and also
to test usage logging.
The logUsage() decorator has been improved to identify method from
function and, if the former, display which class and, via the name
attribute, which instance the method belongs to.
An optional tag for logger date inclusion serves as the first
SystemSettings variable of the module.