Test stubs created and logging file made absolute
A folder for tests and an initial testframework.py file has been stubbed
for the upcoming framework coding work.
A system function called getOptimaCorePath() has also been coded to
return the absolute path to where the module is installed.
This means that logging.ini files no longer need to be copied to the
folder of any application as with Optima TB; there is one logging file
for all applications.
This can be extended later with a try-except clause to have custom
logging files and default to the standard, but it is not a current
priority to handle nonstandard logging.ini files.