Began framework template function
While still considering the best design for a framework file, a
createFrameworkTemplate() function has been introduced in a
framework_io.py file.
So far, it is testing the construction of a population types worksheet
with headers and comments.
Unfortunately, xlsxwriter has shortcomings in formatting comments, but
the current set-up will have to be sufficient.