Hard-coded framework template keys
One way or the other, the framework needs a rigid meta-structure within
which cascade structures can be generically defined.
This meta-structure is arising as hard-coded keys in a FrameworkSettings
class.
In using these definitions, the template framework code has become a
little simpler, as some validation is no longer required to check for
arbitrary keys; the specified keys in FrameworkSettings must exist with
corresponding sections in the configuration file.