Page item descriptors developed
Page items are now part of the configuration file, with the value for
property 'descriptor' serving as the label of an item type for GUI
template creation.
Accordingly, there are extra structures in FrameworkSettings mapping
semantics and page-item keys in reverse.
The only thing left is to set up instruction-updating in the GUI.