Core framework page item loading
The ProjectFramework.importFromFile() method has been developed to load
in core page-items per page.
The current design is that each page is associated with one core item
and subitems that are attached to the core item.
Core items can be compartments, characteristics, parameters, etc.
This will hopefully be made more clear in the design going forward.
Either way, initial construction of specifications dictionaries is now
possible.