Subitem specifications loaded
Each page-item in a framework file must exist as a line of specs data,
including a name.
These lines can be spaced arbitrarily apart in a spreadsheet, with
subitems existing as their own lines of specs data in this space.
In principle, this is a recursive structure with subitems possible for
subitems, although refinement and testing is required.