Excel value extraction made modular
Values in Excel sheets occasionally need to be converted into different formats, depending on specified filters.
The archetypal example is converting 'y' to True and anything else to False (or None, if False is the default).
Function extractExcelSheetValue serves this purpose.
Additionally, output is currently pretty printed in the test script, so as to get better inspection of framework spec dictionaries.