Preliminary extended item-column parser
This commit splits and strips characteristic component strings into lists, although validation is not yet set up.
Given that the intention is to allow a modeller to spread the components over multiple columns and possibly multiple rows, item-column sections have now been set up to expand over subsequent rows that have no item-names, as well as subsequent columns that have no headers.
A framework file to test the limits of modeller input remains to be generated.