Testing spec extraction across rows and columns
The example framework file currently tests the ability of the codebase to import Excel values as scalars or lists across a number of rows and columns.
In general, the last scanned value overwrites any previous value attached to an item and its specification for a particular column header.
However, in the case of a compartment/characteristic components, these are appended to a list.