Semantic overhaul for item type
This commit establishes a clear distinction between individual framework items in a framework file, e.g. a compartment, versus the type of the item, e.g. the compartment type.
The latter revolves around IO interpretation definitions and is stored in FrameworkSettings, while the former is part of a model skeleton and is stored in ProjectFramework.
This distinction between metadata and metametadata is subtle but vital and avoids namespace collisions.