Databook instructions and error propagation
A DatabookInstructions object has been created in the vein of FrameworkTemplateInstructions that specifies how many populations, programs, etc., should be generated in a databook.
Fulfilling these instructions remains to be coded.
Importantly, a fix has been applied to usage logging so that exceptions are stored and raised after timing processes are complete.