Quick breather
Major changes are incoming via the consolidation of workbook input and output for framework files and databooks.
Settings now specify pages and item types specifically, the latter containing specifications for attributes that will immediately describe the layout of ProjectFramework and the new class ProjectData.
Page specifications note table types, which are glorified key objects that detail what kind of tables to construct on a page with reference to which items.
This commit is a functional point in implementational process, but is incomplete.