Started developing semantics dictionary
The base CoreProjectStructure that is derived by both framework and data contains an attribute called spec and an attribute semantics.
The idea behind semantics is to link terms of any type to an item name, what kind of attribute it is, as well as a key list that should allow quick referencing in the specs dictionary.
The first steps have been developed for the DetailColumns table type.