Assumption format hard-coded
Value entry block assumption format is now part of excel.py.
This can be reconsidered later.
Nonetheless, this is necessary so that time-dependent value entry blocks can be parsed without the existence of an "assumption" attribute in item_type_specs.
The issue is now how time-dependent values should actually be stored.