TimeSeries object set up
A rudimentary TimeSeries object that stores a list of lists has been set up.
The first level of this structure corresponds to a time, with 'None' representing a constant assumption, while the second level is split between 't' and provided keys (e.g. program attributes or population disaggregations for parameters and characteristics).
There is no way to add attribute keys after construction at the present time, but lists for new time values are fine.
All access is intended to be done via a set/get method.