GUI restored
Although the GUI is currently having issues running on VS2017, Spyder allows it to operate as it did before all the workbook functionality was overhauled.
In order to achieve this, item types in settings once again take note of superitem types if appropriate.
Also, given that databook settings contains all item types in framework settings, but the user can only dictate the amount of several types, an 'instruction_allowed' tag has been included for item_type_specs in settings as well.
Finally, importing a workbook currently updates the name of a target structure with the workbook path.