Minor terminology change
Data structures must have a static key skeleton, in terms of specifications.
On the other hand, given alternate ways to construct a workbook, it might not make sense to overly hard-code settings with respect to workbook IO.
This commit changes the variable for workbook keys to focus on structure. A specifications initialisation method is also generated for structures.