Probably wrong but first limited result generation
A severely feature-crippled run of the SIR model has been processed, to the point that a results object is created and can be queried, demonstrating characteristics changing over time.
At this point, it is probably time to start implementing feature upgrades, as they will take as much time ironing out bugs in the current workflow.