Content type enforcement
All ContentType objects, which serve as the base class for a hierarchy, now have an enforce_type attribute.
If Spec values are added for any structure with content_type also supplied, values will be converted to this optional enforce_type if possible (except if they are None).
Because most values are supplied in str format, it's not always clear whether type enforcement will work as the user expects. For instance, bool("False") is True and int("-1.0") creates an error.
In the latter case, there is a hard-coded intermediate conversion supplied by float. Either way, type enforcement must be watched for bizarre behaviour.
Nonetheless, the SIR test case is currently working well, with 'datapage_order' values of "-1.0" being convered to -1 ints, subsequently suppressing unnecessary tables being produced in the SIR databook.