Value updating slowly opened up
Junctions are still disabled. Formats are commented out. Special rules do not exist. Programs are ignored or commented out.
However, epi dynamics is beginning to look good, especially with functions now being processed. The SIR test model is progressing.

Note: A 'filter' attribute has been added to ProjectFramework that should be used to list names of objects associated with some key, i.e. for fast referencing.
In this commit, all parameters with functions are noted down in this way.