TimeSeries format preparation
The TimeSeries object, previously storing values as lists of lists, is now stored as a dict of lists.
This loses data contiguity, but is easier to access by key.
Moreover, it allows for special keys like "format" and potentially "unit" in the future.
This is a preface to establishing format-dependent model behaviour.