Basic format setup operational
Although there are likely to be edge cases, rigorous debugging is required and the implementation is too hacky for long-term stability, it is now possible to assign values formats and have them propagate to model iterations.
Part of this involved ensuring that ProjectData had a completeSpecs() method of its own to ensure that all parameters in the framework were defined there, along with relevant formats.
Hopefully, model.py will eventually be set up to query the framework directly about the format of the parameter, but this is a reasonable stand-in solution.