SIR cleaned and tweaked
Got rid of the initial prevalence parameter and made FOI directly dependent on prevalence (i.e. nonlinear).
Also increased duration of infection to slow down recovery rate and increased transmission probability to boost FOI, thus avoiding the immediate disappearance of the epidemic.
Deaths increase at a faster pace early on when there is a pulse of infections, but slows down as people recover.