Minor databook ordering fix
Workbook table parsing is currently rigidly in the order that table objects are generated in the corresponding page specifications.
This rigidity should be corrected, but, for now, tables must be ordered appropriately.
With this fix, the SIR test case works fine, allowing both the test and api script to flow without problem.