Tag for impact parameters introduced
With ease due to the architecture behind workbook construction, an "is_impact" attribute has been added for parameters, intended for reference with progbook construction.
Program types, as a page and item, have been removed for now, with the progbook-focussed implementation of programs intending to spell out functional relations directly.
Spreadsheet updates are intended to be next.