Tightened connection matrices
Standard connection matrices now work in one way alone, and that is to store a list of paired tuples under a specified attribute of a specified item type.
This gets rid of the need to have convoluted 'subkey' logic in specs.
Additionally, because 'read order' solves the order in which pages are processed, workarounds are not required for contact matrices being read before the linking values being defined as names of items.
The responsibility of logical processing order of framework/databook pages is on the developer.