Template table trickiness
Template tables need to be duplicated for all items belonging to an item type at some point prior to workbook construction.
Typically, this is done after framework loading; all template tables in data settings are duplicated.
However, the item type like 'transfer' may not exist until data or databook instructions are actually specified.
Dealing with template duplication at this delayed point will be a design challenge.
In any case, this commit allows the ConnectionMatrix table type to be a standard table or a template type, with slightly different IO operations depending on choice.