Stop detail columns parser at blank row
The import of detail columns (DC) tables can now be stopped at a blank row.
This is good for the transfers page where a DC table is combined with connection matrices.
Accordingly, validation has been introduced at the end of a worksheet parser to exclaim if there are more worksheet rows to parse that aren't covered by tables.
This will need edge case testing but seems to be working at the moment.