Quick commit
Just a comment edit before realising that, perhaps, transfer propagation is complete all the way to the model.
This needs to be tested, along with performance checks.
If this works, transfers are functionally implemented in the new architecture. However, a relation between transfer matrices and TDVE tables needs to be established, to avoid drowning the user in data-entry.
Also, transfer types in the framework need to be implemented in order to get soft-coded default values to work, such as aging.