Enabled table-based item-link referencing
Time-dependent value entry tables can suffer from a lot of rows if iterating of item connections.
This commit modifies cell creation content during databook construction so that alternate content can be produced if an Excel condition is met.
This is specifically used by TDVE item-link tables to see if the cell value associated with a relevant item-link is "y", indicating existence of that connection.
This means that a connection matrix that determines whether those connections exist must store references to its cells during workbook construction, within temp_storage, and this is now achieved.
It is a complicated setup from the developer side, but makes things much easier from the user side when marking transfers.
The user entry blocks need to be updated to allow for conditional diplay before this feature is fully complete.