Fancy Excel footwork
The databook UI for both transfers and interactions were set up, as of the previous commit, so that users would mark a connection as active in a connection matrix and then fill out the row in the subsequent TDVE matrix.
However, particularly in the case of interactions, users are likely to prefer just filling in a matrix of time-independent values.
Using fancy Excel equations, it is now possible to put in values of 'n' in the connection matrix to deactivate the TDVE row, 'y' to activate the row, and a number to activate the row and set the constant default value as that value.
Of course, it is the TDVE tables that are parsed for actual data, so users should be aware by comment or user manual that they are free to provide time-dependent values for these 'binary function' parameters like transfers and interactions, but these will be read in as opposed to the matrix, should there be any value conflict.
In any case, this commit allows several use cases.