Weighted averaging by target population
FOI experienced by a population is usually a contact-average of FOI produced by all other populations (i.e. preference for interaction times those doing the interacting).
In this case, SRC_POP_AVG(parameter, interaction, characteristic) is the appropriate syntax to use for such a parameter.
However, occasionally, it is not the target that contact-weights parameter values of sources, but a source that contact-weights parameter values across targets.
The TGT_POP_AVG function handles this and, as of this commit, appears to work properly.