Import and export alignment bug fixes
An example Gauteng TB progbook has been generated as closely to the original Optima TB programs data as the implementation allows, but a number of fixes had to be applied to progbook import and export functions in order to allow the TB version to be imported like the SIR one.
A test case must now be generated to run an example simulation, but propagation at least seems to function.