Don't add TDC names to the reference dictionary
There is a potential edge case if the TDC name duplicates a population name. This might still be permissible elsewhere in the code. However, adding the TDC name to the reference dictionary would result in an incorrect reference if it was duplicated, but is also unnecessary because there shouldn't be any references to the TDC name